+++ index.html
@@ -229,5 +229,5 @@
 
 </body>
 <script src="js/function.js"></script>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 </html>
